concat 2 string

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -26,7 +26,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	for (lenS1 = 0; s1[lenS1]; lenS1++)
 	;
 
-	totalLength = (unsigned int)lenS1 + n + 1;
+	totalLength = lenS1 + n + 1;
 
 	p = malloc(sizeof(char) * totalLength);
